services/meter/requirements.txt
wheel==0.35.1
pkg-resources==0.0.0
python-dotenv==0.15.0
pika==1.1.0
typing-extensions==3.7.4.3
six==1.15.0
tenacity==6.2.0
